@@ -15,17 +15,15 @@ import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlC
 import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_RESEARCH_KEY;
 import static com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlConstants.FIELD_SLIMEFUN_ID;
 
-import city.norain.slimefun4.utils.SimpleTimer;
+import city.norain.slimefun4.timings.entry.SQLEntry;
 import com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlCommonAdapter;
 import com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.SqlUtils;
 import com.xzavier0722.mc.plugin.slimefun4.storage.common.DataScope;
 import com.xzavier0722.mc.plugin.slimefun4.storage.common.DataType;
 import com.xzavier0722.mc.plugin.slimefun4.storage.common.RecordKey;
 import com.xzavier0722.mc.plugin.slimefun4.storage.common.RecordSet;
-import io.github.thebusybiscuit.slimefun4.core.debug.Debug;
-import io.github.thebusybiscuit.slimefun4.core.debug.TestCase;
+import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
 import java.sql.SQLException;
-import java.time.Duration;
 import java.util.List;
 
 public class SqliteAdapter extends SqlCommonAdapter<SqliteConfig> {
@@ -320,15 +318,15 @@ public class SqliteAdapter extends SqlCommonAdapter<SqliteConfig> {
     }
 
     private synchronized int executeUpdate(String sql) {
-        var timer = new SimpleTimer();
+        var entry = new SQLEntry(sql);
+        Slimefun.getSQLProfiler().recordEntry(entry);
+
         try (var conn = ds.getConnection()) {
             return SqlUtils.execUpdate(conn, sql);
         } catch (SQLException e) {
             throw new IllegalStateException("An exception thrown while executing sql: " + sql, e);
         } finally {
-            if (timer.isTimeout(Duration.ofSeconds(2))) { // FIXME: hardcode slow sql check duration
-                Debug.log(TestCase.DATABASE, "Detected slow sql costs {}, sql: {}", timer.durationStr(), sql);
-            }
+            Slimefun.getSQLProfiler().finishEntry(entry);
         }
     }
 }
