@@ -1,5 +1,6 @@
 package com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlite;
 
+import city.norain.slimefun4.SlimefunExtended;
 import com.xzavier0722.mc.plugin.slimefun4.storage.adapter.sqlcommon.ISqlCommonConfig;
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
@@ -11,7 +12,14 @@ public record SqliteConfig(String path, int maxConnection) implements ISqlCommon
         config.setJdbcUrl(jdbcUrl());
         config.setPoolName("SlimefunHikariPool");
         config.setMaximumPoolSize(maxConnection);
-        config.setLeakDetectionThreshold(3000);
+
+        /* ===================
+         * DATABASE DEBUG MODE
+         * ===================
+         */
+        if (SlimefunExtended.isDatabaseDebugMode()) {
+            config.setLeakDetectionThreshold(5000);
+        }
 
         return new HikariDataSource(config);
     }
