@@ -11,9 +11,7 @@ import org.bukkit.Server;
  *
  * @author TheBusyBiscuit
  * @author Walshy
- *
  * @see Slimefun
- *
  */
 public enum MinecraftVersion {
 
@@ -48,6 +46,18 @@ public enum MinecraftVersion {
     MINECRAFT_1_20(20, "1.20.x"),
 
     /**
+     * This constant represents Minecraft (Java Edition) Version 1.20.5
+     * ("The Armored Paws Update")
+     */
+    MINECRAFT_1_20_5(20, 5, "1.20.5+"),
+
+    /**
+     * This constant represents Minecraft (Java Edition) Version 1.21
+     * ("The Tricky Trials Update")
+     */
+    MINECRAFT_1_21(21, "1.21.x"),
+
+    /**
      * This constant represents an exceptional state in which we were unable
      * to identify the Minecraft Version we are using
      */
@@ -62,20 +72,36 @@ public enum MinecraftVersion {
     private final String name;
     private final boolean virtual;
     private final int majorVersion;
+    private final int minorVersion;
 
     /**
      * This constructs a new {@link MinecraftVersion} with the given name.
      * This constructor forces the {@link MinecraftVersion} to be real.
      * It must be a real version of Minecraft.
      *
-     * @param majorVersion
-     *            The major version of minecraft as an {@link Integer}
-     * @param name
-     *            The display name of this {@link MinecraftVersion}
+     * @param majorVersion The major version of minecraft as an {@link Integer}
+     * @param name         The display name of this {@link MinecraftVersion}
      */
     MinecraftVersion(int majorVersion, @Nonnull String name) {
         this.name = name;
         this.majorVersion = majorVersion;
+        this.minorVersion = -1;
+        this.virtual = false;
+    }
+
+    /**
+     * This constructs a new {@link MinecraftVersion} with the given name.
+     * This constructor forces the {@link MinecraftVersion} to be real.
+     * It must be a real version of Minecraft.
+     *
+     * @param majorVersion The major (minor in semver, major in MC land) version of minecraft as an {@link Integer}
+     * @param minor        The minor (patch in semver, minor in MC land) version of minecraft as an {@link Integer}
+     * @param name         The display name of this {@link MinecraftVersion}
+     */
+    MinecraftVersion(int majorVersion, int minor, @Nonnull String name) {
+        this.name = name;
+        this.majorVersion = majorVersion;
+        this.minorVersion = minor;
         this.virtual = false;
     }
 
@@ -84,14 +110,13 @@ public enum MinecraftVersion {
      * A virtual {@link MinecraftVersion} (unknown or unit test) is not an actual
      * version of Minecraft but rather a state of the {@link Server} software.
      *
-     * @param name
-     *            The display name of this {@link MinecraftVersion}
-     * @param virtual
-     *            Whether this {@link MinecraftVersion} is virtual
+     * @param name    The display name of this {@link MinecraftVersion}
+     * @param virtual Whether this {@link MinecraftVersion} is virtual
      */
     MinecraftVersion(@Nonnull String name, boolean virtual) {
         this.name = name;
         this.majorVersion = 0;
+        this.minorVersion = -1;
         this.virtual = virtual;
     }
 
@@ -126,13 +151,42 @@ public enum MinecraftVersion {
      * <p>
      * Example: {@literal "1.13"} returns {@literal 13}
      *
-     * @param minecraftVersion
-     *            The {@link Integer} version to match
-     *
+     * @param minecraftVersion The {@link Integer} version to match
      * @return Whether this {@link MinecraftVersion} matches the specified version id
      */
     public boolean isMinecraftVersion(int minecraftVersion) {
-        return !isVirtual() && this.majorVersion == minecraftVersion;
+        return this.isMinecraftVersion(minecraftVersion, -1);
+    }
+
+    /**
+     * This tests if the given minecraft version matches with this
+     * {@link MinecraftVersion}.
+     * <p>
+     * You can obtain the version number by doing {@link PaperLib#getMinecraftVersion()}.
+     * It is equivalent to the "major" version<br />
+     * You can obtain the patch version by doing {@link PaperLib#getMinecraftPatchVersion()}.
+     * It is equivalent to the "minor" version
+     * <p>
+     * Example: {@literal "1.13"} returns {@literal 13}<br />
+     * Exampe: {@literal "1.13.2"} returns {@literal 13_2}
+     *
+     * @param minecraftVersion The {@link Integer} version to match
+     * @return Whether this {@link MinecraftVersion} matches the specified version id
+     */
+    public boolean isMinecraftVersion(int minecraftVersion, int patchVersion) {
+        if (isVirtual()) {
+            return false;
+        }
+
+        if (this.majorVersion != minecraftVersion) {
+            return false;
+        }
+
+        if (this.majorVersion == 20) {
+            return this.minorVersion == -1 ? patchVersion < 5 : patchVersion >= this.minorVersion;
+        } else {
+            return this.minorVersion == -1 || patchVersion >= this.minorVersion;
+        }
     }
 
     /**
@@ -141,9 +195,7 @@ public enum MinecraftVersion {
      *
      * An unknown version will default to {@literal false}.
      *
-     * @param version
-     *            The {@link MinecraftVersion} to compare
-     *
+     * @param version The {@link MinecraftVersion} to compare
      * @return Whether this {@link MinecraftVersion} is newer or equal to the given {@link MinecraftVersion}
      */
     public boolean isAtLeast(@Nonnull MinecraftVersion version) {
@@ -176,9 +228,7 @@ public enum MinecraftVersion {
      *
      * An unknown version will default to {@literal true}.
      *
-     * @param version
-     *            The {@link MinecraftVersion} to compare
-     *
+     * @param version The {@link MinecraftVersion} to compare
      * @return Whether this {@link MinecraftVersion} is older than the given one
      */
     public boolean isBefore(@Nonnull MinecraftVersion version) {
