@@ -3,6 +3,7 @@ package io.github.thebusybiscuit.slimefun4.core.attributes;
 import io.github.bakedlibs.dough.config.Config;
 import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItem;
 import io.github.thebusybiscuit.slimefun4.utils.UnbreakingAlgorithm;
+import io.github.thebusybiscuit.slimefun4.utils.compatibility.VersionedEnchantment;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import org.bukkit.Sound;
@@ -49,7 +50,7 @@ public interface DamageableItem extends ItemAttribute {
      */
     default void damageItem(@Nonnull Player p, @Nullable ItemStack item) {
         if (isDamageable() && item != null && !item.getType().isAir() && item.getAmount() > 0) {
-            int unbreakingLevel = item.getEnchantmentLevel(Enchantment.DURABILITY);
+            int unbreakingLevel = item.getEnchantmentLevel(VersionedEnchantment.UNBREAKING);
 
             if (evaluateUnbreakingEnchantment(unbreakingLevel)) {
                 return;
