@@ -24,7 +24,9 @@ public enum TestCase {
 
     UTILS,
 
-    ENERGYNET;
+    ENERGYNET,
+
+    ANALYTICS;
 
     TestCase() {}
 
