@@ -1,13 +1,13 @@
 package io.github.thebusybiscuit.slimefun4.core.services.github;
 
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
 import io.github.thebusybiscuit.slimefun4.utils.NumberUtils;
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.Map;
 import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
-import kong.unirest.JsonNode;
-import kong.unirest.json.JSONObject;
 
 class GitHubActivityConnector extends GitHubConnector {
 
@@ -20,11 +20,12 @@ class GitHubActivityConnector extends GitHubConnector {
     }
 
     @Override
-    public void onSuccess(@Nonnull JsonNode response) {
-        JSONObject object = response.getObject();
-        int forks = object.getInt("forks");
-        int stars = object.getInt("stargazers_count");
-        LocalDateTime lastPush = NumberUtils.parseGitHubDate(object.getString("pushed_at"));
+    public void onSuccess(@Nonnull JsonElement response) {
+        JsonObject object = response.getAsJsonObject();
+        int forks = object.get("forks").getAsInt();
+        int stars = object.get("stargazers_count").getAsInt();
+        LocalDateTime lastPush =
+                NumberUtils.parseGitHubDate(object.get("pushed_at").getAsString());
 
         callback.accept(forks, stars, lastPush);
     }
