@@ -1,14 +1,14 @@
 package io.github.thebusybiscuit.slimefun4.core.services.github;
 
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
 import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.logging.Level;
 import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
-import kong.unirest.JsonNode;
-import kong.unirest.json.JSONArray;
-import kong.unirest.json.JSONObject;
 
 class GitHubIssuesConnector extends GitHubConnector {
 
@@ -21,15 +21,15 @@ class GitHubIssuesConnector extends GitHubConnector {
     }
 
     @Override
-    public void onSuccess(@Nonnull JsonNode response) {
-        if (response.isArray()) {
-            JSONArray array = response.getArray();
+    public void onSuccess(@Nonnull JsonElement response) {
+        if (response.isJsonArray()) {
+            JsonArray array = response.getAsJsonArray();
 
             int issues = 0;
             int pullRequests = 0;
 
-            for (int i = 0; i < array.length(); i++) {
-                JSONObject obj = array.getJSONObject(i);
+            for (JsonElement element : array) {
+                JsonObject obj = element.getAsJsonObject();
 
                 if (obj.has("pull_request")) {
                     pullRequests++;
