@@ -1,5 +1,6 @@
 package io.github.thebusybiscuit.slimefun4.core.services.profiler;
 
+import com.google.common.util.concurrent.AtomicDouble;
 import io.github.thebusybiscuit.slimefun4.api.SlimefunAddon;
 import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItem;
 import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
@@ -85,6 +86,8 @@ public class SlimefunProfiler {
 
     private final AtomicLong totalMsTicked = new AtomicLong();
     private final AtomicInteger ticksPassed = new AtomicInteger();
+    private final AtomicLong totalNsTicked = new AtomicLong();
+    private final AtomicDouble averageTimingsPerMachine = new AtomicDouble();
 
     /**
      * This method terminates the {@link SlimefunProfiler}.
@@ -203,8 +206,7 @@ public class SlimefunProfiler {
                     while (iterator.hasNext()) {
                         iterator.next()
                                 .sendMessage("Your timings report has timed out, we were still waiting for "
-                                        + queued.get()
-                                        + " samples to be collected :/");
+                                        + queued.get() + " samples to be collected :/");
                         iterator.remove();
                     }
 
@@ -223,11 +225,15 @@ public class SlimefunProfiler {
 
         totalElapsedTime = timings.values().stream().mapToLong(Long::longValue).sum();
 
+        averageTimingsPerMachine.getAndSet(
+                timings.values().stream().mapToLong(Long::longValue).average().orElse(0));
+
         /*
          * We log how many milliseconds have been ticked, and how many ticks have passed
          * This is so when bStats requests the average timings, they're super quick to figure out
          */
         totalMsTicked.addAndGet(TimeUnit.NANOSECONDS.toMillis(totalElapsedTime));
+        totalNsTicked.addAndGet(totalElapsedTime);
         ticksPassed.incrementAndGet();
 
         if (!requests.isEmpty()) {
@@ -417,5 +423,27 @@ public class SlimefunProfiler {
         ticksPassed.set(0);
 
         return l;
+    }
+
+    /**
+     * Get and reset the average nanosecond timing for this {@link SlimefunProfiler}.
+     *
+     * @return The average nanosecond timing for this {@link SlimefunProfiler}.
+     */
+    public double getAndResetAverageNanosecondTimings() {
+        long l = totalNsTicked.get() / ticksPassed.get();
+        totalNsTicked.set(0);
+        ticksPassed.set(0);
+
+        return l;
+    }
+
+    /**
+     * Get and reset the average millisecond timing for each machine.
+     *
+     * @return The average millisecond timing for each machine.
+     */
+    public double getAverageTimingsPerMachine() {
+        return averageTimingsPerMachine.getAndSet(0);
     }
 }
