@@ -9,12 +9,14 @@ import io.github.thebusybiscuit.slimefun4.core.handlers.EntityInteractHandler;
 import io.github.thebusybiscuit.slimefun4.core.handlers.ItemUseHandler;
 import io.github.thebusybiscuit.slimefun4.core.handlers.ToolUseHandler;
 import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
+import io.github.thebusybiscuit.slimefun4.utils.compatibility.VersionedEntityType;
 import java.util.ArrayList;
 import java.util.List;
 import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 import org.bukkit.ChatColor;
 import org.bukkit.NamespacedKey;
+import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.persistence.PersistentDataType;
@@ -116,15 +118,12 @@ public class MultiTool extends SlimefunItem implements Rechargeable {
     private EntityInteractHandler getEntityInteractionHandler() {
         return (e, item, offhand) -> {
             // Fixes #2217 - Prevent them from being used to shear entities
-            switch (e.getRightClicked().getType()) {
-                case MUSHROOM_COW:
-                case SHEEP:
-                case SNOWMAN:
-                    Slimefun.getLocalization().sendMessage(e.getPlayer(), "messages.multi-tool.not-shears");
-                    e.setCancelled(true);
-                    break;
-                default:
-                    break;
+            EntityType type = e.getRightClicked().getType();
+            if (type == VersionedEntityType.MOOSHROOM
+                    || type == VersionedEntityType.SNOW_GOLEM
+                    || type == EntityType.SHEEP) {
+                Slimefun.getLocalization().sendMessage(e.getPlayer(), "messages.multi-tool.not-shears");
+                e.setCancelled(true);
             }
         };
     }
