@@ -6,9 +6,9 @@ import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItemStack;
 import io.github.thebusybiscuit.slimefun4.api.recipes.RecipeType;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunItems;
 import io.github.thebusybiscuit.slimefun4.utils.SlimefunUtils;
+import io.github.thebusybiscuit.slimefun4.utils.compatibility.VersionedParticle;
 import io.github.thebusybiscuit.slimefun4.utils.itemstack.ItemStackWrapper;
 import me.mrCookieSlime.Slimefun.api.inventory.BlockMenu;
-import org.bukkit.Particle;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Ageable;
 import org.bukkit.entity.Entity;
@@ -55,7 +55,12 @@ public class AnimalGrowthAccelerator extends AbstractGrowthAccelerator {
 
                     n.getWorld()
                             .spawnParticle(
-                                    Particle.VILLAGER_HAPPY, ((LivingEntity) n).getEyeLocation(), 8, 0.2F, 0.2F, 0.2F);
+                                    VersionedParticle.HAPPY_VILLAGER,
+                                    ((LivingEntity) n).getEyeLocation(),
+                                    8,
+                                    0.2F,
+                                    0.2F,
+                                    0.2F);
                     return;
                 }
             }
