@@ -8,6 +8,7 @@ import io.github.thebusybiscuit.slimefun4.core.handlers.ItemDropHandler;
 import io.github.thebusybiscuit.slimefun4.core.services.sounds.SoundEffect;
 import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
 import io.github.thebusybiscuit.slimefun4.implementation.items.SimpleSlimefunItem;
+import io.github.thebusybiscuit.slimefun4.utils.compatibility.VersionedParticle;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.EnumMap;
@@ -20,7 +21,6 @@ import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 import org.bukkit.Location;
 import org.bukkit.Material;
-import org.bukkit.Particle;
 import org.bukkit.SoundCategory;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.entity.Entity;
@@ -50,6 +50,10 @@ public class EnchantmentRune extends SimpleSlimefunItem<ItemDropHandler> {
             List<Enchantment> enchantments = new ArrayList<>();
 
             for (Enchantment enchantment : Enchantment.values()) {
+                if (!mat.isItem()) {
+                    continue;
+                }
+
                 if (enchantment.equals(Enchantment.BINDING_CURSE) || enchantment.equals(Enchantment.VANISHING_CURSE)) {
                     continue;
                 }
@@ -142,7 +146,7 @@ public class EnchantmentRune extends SimpleSlimefunItem<ItemDropHandler> {
                             // Being sure entities are still valid and not picked up or whatsoever.
                             if (rune.isValid() && item.isValid() && itemStack.getAmount() == 1) {
 
-                                l.getWorld().spawnParticle(Particle.CRIT_MAGIC, l, 1);
+                                l.getWorld().spawnParticle(VersionedParticle.ENCHANTED_HIT, l, 1);
                                 SoundEffect.ENCHANTMENT_RUNE_ADD_ENCHANT_SOUND.playAt(l, SoundCategory.PLAYERS);
 
                                 item.remove();
