@@ -17,6 +17,7 @@ import io.github.thebusybiscuit.slimefun4.core.handlers.BlockBreakHandler;
 import io.github.thebusybiscuit.slimefun4.core.handlers.BlockPlaceHandler;
 import io.github.thebusybiscuit.slimefun4.core.handlers.ToolUseHandler;
 import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
+import io.github.thebusybiscuit.slimefun4.utils.compatibility.VersionedEnchantment;
 import io.github.thebusybiscuit.slimefun4.utils.tags.SlimefunTag;
 import java.util.ArrayList;
 import java.util.List;
@@ -27,6 +28,7 @@ import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import org.bukkit.Bukkit;
 import org.bukkit.GameMode;
+import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
@@ -128,11 +130,11 @@ public class BlockListener implements Listener {
                     } else if (sfItem instanceof NotRotatable notRotatable) {
                         rotation = notRotatable.getRotation();
                     } else if (sfItem instanceof NotCardinallyRotatable notRotatable) {
-                        rotation = notRotatable.getRotation(
-                                e.getPlayer().getLocation().getYaw());
+                        rotation = notRotatable.getRotation(Location.normalizeYaw(
+                                e.getPlayer().getLocation().getYaw()));
                     } else if (sfItem instanceof NotDiagonallyRotatable notRotatable) {
-                        rotation = notRotatable.getRotation(
-                                e.getPlayer().getLocation().getYaw());
+                        rotation = notRotatable.getRotation(Location.normalizeYaw(
+                                e.getPlayer().getLocation().getYaw()));
                     }
 
                     if (rotation != null) {
@@ -174,11 +176,11 @@ public class BlockListener implements Listener {
         var heldItem = e.getPlayer().getInventory().getItemInMainHand();
         var block = e.getBlock();
         var blockData = StorageCacheUtils.getBlock(block.getLocation());
+        var sfItem = blockData == null ? null : SlimefunItem.getById(blockData.getSfId());
 
         // If there is a Slimefun Block here, call our BreakEvent and, if cancelled, cancel this event
         // and return
         if (blockData != null) {
-            var sfItem = SlimefunItem.getById(blockData.getSfId());
             SlimefunBlockBreakEvent breakEvent =
                     new SlimefunBlockBreakEvent(e.getPlayer(), heldItem, e.getBlock(), sfItem);
             Bukkit.getPluginManager().callEvent(breakEvent);
@@ -202,7 +204,7 @@ public class BlockListener implements Listener {
             checkForSensitiveBlockAbove(e.getPlayer(), e.getBlock(), heldItem);
 
             if (blockData == null || blockData.isPendingRemove()) {
-                dropItems(e, drops);
+                dropItems(e, heldItem, block, sfItem, drops);
                 return;
             }
 
@@ -221,7 +223,7 @@ public class BlockListener implements Listener {
                                 return;
                             }
                             e.setDropItems(true);
-                            dropItems(e, drops);
+                            dropItems(e, heldItem, block, sfItem, drops);
                         },
                         true);
                 return;
@@ -231,10 +233,10 @@ public class BlockListener implements Listener {
             if (e.isCancelled()) {
                 blockData.setPendingRemove(false);
             }
-            dropItems(e, drops);
+            dropItems(e, heldItem, block, sfItem, drops);
 
             // Checks for vanilla sensitive blocks everywhere
-            checkForSensitiveBlocks(e.getBlock(), 0, e.isDropItems());
+            // checkForSensitiveBlocks(e.getBlock(), 0, e.isDropItems());
         }
     }
 
@@ -269,12 +271,13 @@ public class BlockListener implements Listener {
     }
 
     @ParametersAreNonnullByDefault
-    private void dropItems(BlockBreakEvent e, List<ItemStack> drops) {
+    private void dropItems(
+            BlockBreakEvent e, ItemStack item, Block block, @Nullable SlimefunItem sfBlock, List<ItemStack> drops) {
         if (!drops.isEmpty()) {
             // TODO: properly support loading inventories within unit tests
             if (!Slimefun.instance().isUnitTest()) {
                 // Notify plugins like CoreProtect
-                Slimefun.getProtectionManager().logAction(e.getPlayer(), e.getBlock(), Interaction.BREAK_BLOCK);
+                Slimefun.getProtectionManager().logAction(e.getPlayer(), block, Interaction.BREAK_BLOCK);
             }
 
             // Fixes #2560
@@ -282,14 +285,18 @@ public class BlockListener implements Listener {
                 // Disable normal block drops
                 e.setDropItems(false);
 
+                // Fixes #4051
+                if (sfBlock == null) {
+                    block.breakNaturally(item);
+                }
+
+                // The list only contains other drops, not those from the block itself, so we still need to handle those
                 for (ItemStack drop : drops) {
                     // Prevent null or air from being dropped
                     if (drop != null && drop.getType() != Material.AIR) {
                         if (e.getPlayer().getGameMode() != GameMode.CREATIVE
                                 || Slimefun.getCfg().getBoolean("options.drop-block-creative")) {
-                            e.getBlock()
-                                    .getWorld()
-                                    .dropItemNaturally(e.getBlock().getLocation(), drop);
+                            block.getWorld().dropItemNaturally(block.getLocation(), drop);
                         }
                     }
                 }
@@ -328,7 +335,7 @@ public class BlockListener implements Listener {
                     sfItem.callItemHandler(
                             BlockBreakHandler.class, handler -> handler.onPlayerBreak(dummyEvent, item, drops));
                     controller.removeBlock(loc);
-                    dropItems(dummyEvent, drops);
+                    dropItems(dummyEvent, item, block, sfItem, drops);
                 } else {
                     blockData.setPendingRemove(true);
                     controller.loadBlockDataAsync(blockData, new IAsyncReadCallback<>() {
@@ -342,7 +349,7 @@ public class BlockListener implements Listener {
                             sfItem.callItemHandler(
                                     BlockBreakHandler.class, handler -> handler.onPlayerBreak(dummyEvent, item, drops));
                             controller.removeBlock(loc);
-                            dropItems(dummyEvent, drops);
+                            dropItems(dummyEvent, item, block, sfItem, drops);
                         }
                     });
                 }
@@ -418,7 +425,7 @@ public class BlockListener implements Listener {
              * directly and reuse it.
              */
             ItemMeta meta = item.getItemMeta();
-            int fortuneLevel = meta.getEnchantLevel(Enchantment.LOOT_BONUS_BLOCKS);
+            int fortuneLevel = meta.getEnchantLevel(VersionedEnchantment.FORTUNE);
 
             if (fortuneLevel > 0 && !meta.hasEnchant(Enchantment.SILK_TOUCH)) {
                 Random random = ThreadLocalRandom.current();
