@@ -27,8 +27,10 @@ public class JoinListener implements Listener {
 
     @EventHandler
     public void onJoin(@Nonnull PlayerJoinEvent e) {
-        PlayerProfile.get(e.getPlayer(), playerProfile -> {
-            final ItemStack[] armorContents = e.getPlayer().getInventory().getArmorContents();
+        final var p = e.getPlayer();
+
+        PlayerProfile.get(p, playerProfile -> {
+            final ItemStack[] armorContents = p.getInventory().getArmorContents();
             final HashedArmorpiece[] hashedArmorpieces = playerProfile.getArmor();
             for (int i = 0; i < 4; i++) {
                 final ItemStack armorPiece = armorContents[i];
