@@ -4,9 +4,11 @@ import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
 import io.github.thebusybiscuit.slimefun4.implementation.tasks.armor.RadiationTask;
 import io.github.thebusybiscuit.slimefun4.utils.RadiationUtils;
 import javax.annotation.Nonnull;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.PlayerDeathEvent;
+import org.bukkit.event.player.PlayerJoinEvent;
 
 /**
  * {@link RadioactivityListener} handles radioactivity level resets
@@ -22,7 +24,16 @@ public class RadioactivityListener implements Listener {
 
     @EventHandler
     public void onPlayerDeath(@Nonnull PlayerDeathEvent e) {
-        RadiationUtils.clearExposure(e.getEntity());
-        RadiationTask.addGracePeriod(e.getEntity());
+        addGracePeriod(e.getEntity());
+    }
+
+    @EventHandler
+    public void onPlayerJoin(@Nonnull PlayerJoinEvent e) {
+        addGracePeriod(e.getPlayer());
+    }
+
+    private void addGracePeriod(@Nonnull Player entity) {
+        RadiationUtils.clearExposure(entity);
+        RadiationTask.addGracePeriod(entity);
     }
 }
