@@ -1,5 +1,6 @@
 package io.github.thebusybiscuit.slimefun4.implementation.listeners;
 
+import city.norain.slimefun4.compatibillty.VersionedEvent;
 import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItem;
 import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
 import io.github.thebusybiscuit.slimefun4.implementation.items.VanillaItem;
@@ -30,8 +31,8 @@ public class VillagerTradingListener implements Listener {
 
     @EventHandler(ignoreCancelled = true)
     public void onPreTrade(InventoryClickEvent e) {
-        Inventory clickedInventory = e.getClickedInventory();
-        Inventory topInventory = e.getView().getTopInventory();
+        Inventory clickedInventory = VersionedEvent.getClickedInventory(e);
+        Inventory topInventory = VersionedEvent.getTopInventory(e);
 
         if (clickedInventory != null && topInventory.getType() == InventoryType.MERCHANT) {
             if (e.getAction() == InventoryAction.HOTBAR_SWAP) {
