@@ -9,6 +9,7 @@ import io.github.thebusybiscuit.slimefun4.core.attributes.Radioactive;
 import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
 import io.github.thebusybiscuit.slimefun4.implementation.items.armor.SlimefunArmorPiece;
 import io.github.thebusybiscuit.slimefun4.implementation.items.electric.gadgets.SolarHelmet;
+import io.github.thebusybiscuit.slimefun4.utils.compatibility.VersionedPotionEffectType;
 import io.github.thebusybiscuit.slimefun4.utils.itemstack.ItemStackWrapper;
 import java.util.Collections;
 import java.util.HashSet;
@@ -50,10 +51,10 @@ public class ArmorTask implements Runnable {
         Set<PotionEffect> effects = new HashSet<>();
         effects.add(new PotionEffect(PotionEffectType.WITHER, 400, 2));
         effects.add(new PotionEffect(PotionEffectType.BLINDNESS, 400, 3));
-        effects.add(new PotionEffect(PotionEffectType.CONFUSION, 400, 3));
+        effects.add(new PotionEffect(VersionedPotionEffectType.NAUSEA, 400, 3));
         effects.add(new PotionEffect(PotionEffectType.WEAKNESS, 400, 2));
-        effects.add(new PotionEffect(PotionEffectType.SLOW, 400, 1));
-        effects.add(new PotionEffect(PotionEffectType.SLOW_DIGGING, 400, 1));
+        effects.add(new PotionEffect(VersionedPotionEffectType.SLOWNESS, 400, 1));
+        effects.add(new PotionEffect(VersionedPotionEffectType.MINING_FATIGUE, 400, 1));
         radiationEffects = Collections.unmodifiableSet(effects);
     }
 
