@@ -1,5 +1,7 @@
 package io.github.thebusybiscuit.slimefun4.utils;
 
+import io.github.thebusybiscuit.slimefun4.api.MinecraftVersion;
+import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
 import javax.annotation.Nonnull;
@@ -14,12 +16,14 @@ import org.bukkit.entity.Firework;
 import org.bukkit.inventory.meta.FireworkMeta;
 
 /**
- * This is a simple utility classs for spawning random and colorful {@link Firework} rockets.
+ * This is a simple utility class for spawning random and colorful {@link Firework} rockets.
  *
  * @author TheBusyBiscuit
+ *
  */
 public final class FireworkUtils {
 
+    // @formatter:off
     private static final Color[] COLORS = {
         Color.AQUA,
         Color.BLACK,
@@ -39,6 +43,17 @@ public final class FireworkUtils {
         Color.WHITE,
         Color.YELLOW
     };
+    // @formatter:on
+
+    private static final EntityType firework;
+
+    static {
+        if (Slimefun.getMinecraftVersion().isAtLeast(MinecraftVersion.MINECRAFT_1_20_5)) {
+            firework = EntityType.FIREWORK_ROCKET;
+        } else {
+            firework = EntityType.valueOf("FIREWORK");
+        }
+    }
 
     private FireworkUtils() {}
 
@@ -46,9 +61,8 @@ public final class FireworkUtils {
         createFirework(l, color);
     }
 
-    @Nonnull
-    public static Firework createFirework(@Nonnull Location l, @Nonnull Color color) {
-        Firework fw = (Firework) l.getWorld().spawnEntity(l, EntityType.FIREWORK);
+    public static @Nonnull Firework createFirework(@Nonnull Location l, @Nonnull Color color) {
+        Firework fw = (Firework) l.getWorld().spawnEntity(l, firework);
         FireworkMeta meta = fw.getFireworkMeta();
 
         meta.setDisplayName(ChatColor.GREEN + "Slimefun Research");
@@ -72,18 +86,33 @@ public final class FireworkUtils {
         }
     }
 
-    @Nonnull
-    public static FireworkEffect getRandomEffect(@Nonnull Random random, @Nonnull Color color) {
+    /**
+     * This returns a randomized {@link FireworkEffect} using the given {@link Color}.
+     *
+     * @param random
+     *            The {@link Random} instance to use
+     * @param color
+     *            The {@link Color} of this {@link FireworkEffect}
+     *
+     * @return A randomized {@link FireworkEffect}
+     */
+    public static @Nonnull FireworkEffect getRandomEffect(@Nonnull Random random, @Nonnull Color color) {
+        // @formatter:off
         return FireworkEffect.builder()
                 .flicker(random.nextBoolean())
                 .withColor(color)
                 .with(random.nextBoolean() ? Type.BALL : Type.BALL_LARGE)
                 .trail(random.nextBoolean())
                 .build();
+        // @formatter:on
     }
 
-    @Nonnull
-    private static Color getRandomColor() {
+    /**
+     * This returns a random {@link Color} for our {@link FireworkEffect}.
+     *
+     * @return A random {@link Color}
+     */
+    private static @Nonnull Color getRandomColor() {
         return COLORS[ThreadLocalRandom.current().nextInt(COLORS.length)];
     }
 }
