@@ -28,6 +28,12 @@ public class ChestMenu extends SlimefunInventoryHolder {
     private boolean emptyClickable;
     private String title;
     private List<ItemStack> items;
+    /**
+     * Size of chestmenu
+     * Warning: it DOES NOT present actual size of its inventory!
+     */
+    private int size = -1;
+
     private Map<Integer, MenuClickHandler> handlers;
     private MenuOpeningHandler open;
     private MenuCloseHandler close;
@@ -50,6 +56,11 @@ public class ChestMenu extends SlimefunInventoryHolder {
         this.open = p -> {};
         this.close = p -> {};
         this.playerclick = (p, slot, item, action) -> isPlayerInventoryClickable();
+    }
+
+    public ChestMenu(String title, int size) {
+        this(title);
+        setSize(size);
     }
 
     /**
@@ -116,14 +127,17 @@ public class ChestMenu extends SlimefunInventoryHolder {
      * @return The ChestMenu Instance
      */
     public ChestMenu addItem(int slot, ItemStack item) {
-        final int size = this.items.size();
-        if (size > slot) this.items.set(slot, item);
-        else {
-            for (int i = 0; i < slot - size; i++) {
-                this.items.add(null);
-            }
-            this.items.add(item);
-        }
+        /**
+         * // do shallow copy due to Paper ItemStack system change
+         * // See also: https://github.com/PaperMC/Paper/pull/10852
+         * ItemStack clone = item == null ? null : new ItemStack(item.getType(), item.getAmount());
+         *
+         * if (clone != null && item.hasItemMeta()) {
+         * clone.setItemMeta(item.getItemMeta());
+         * }*/
+        setSize((int) (Math.max(getSize(), Math.ceil((slot + 1) / 9d) * 9)));
+        this.items.set(slot, item);
+        this.inventory.setItem(slot, item);
         return this;
     }
 
@@ -150,6 +164,9 @@ public class ChestMenu extends SlimefunInventoryHolder {
      */
     public ItemStack getItemInSlot(int slot) {
         setup();
+        if (items.size() - 1 < slot) {
+            addItem(slot, null);
+        }
         return this.inventory.getItem(slot);
     }
 
@@ -225,7 +242,7 @@ public class ChestMenu extends SlimefunInventoryHolder {
 
     private void setup() {
         if (this.inventory != null) return;
-        this.inventory = Bukkit.createInventory(this, ((int) Math.ceil(this.items.size() / 9F)) * 9, title);
+        this.inventory = Bukkit.createInventory(this, getSize(), title);
         for (int i = 0; i < this.items.size(); i++) {
             this.inventory.setItem(i, this.items.get(i));
         }
@@ -235,8 +252,15 @@ public class ChestMenu extends SlimefunInventoryHolder {
      * Resets this ChestMenu to a Point BEFORE the User interacted with it
      */
     public void reset(boolean update) {
-        if (update) this.inventory.clear();
-        else this.inventory = Bukkit.createInventory(this, ((int) Math.ceil(this.items.size() / 9F)) * 9, title);
+        if (this.inventory == null || this.inventory.getSize() != getSize())
+            this.inventory = Bukkit.createInventory(this, getSize(), title);
+
+        if (update) {
+            this.inventory.clear();
+        } else {
+            this.inventory = Bukkit.createInventory(this, getSize(), title);
+        }
+
         for (int i = 0; i < this.items.size(); i++) {
             this.inventory.setItem(i, this.items.get(i));
         }
@@ -313,6 +337,43 @@ public class ChestMenu extends SlimefunInventoryHolder {
      */
     public Inventory toInventory() {
         return this.inventory;
+    }
+
+    public int getSize() {
+        return isSizeAutomaticallyInferred() ? Math.max(9, (int) Math.ceil(this.items.size() / 9F) * 9) : size;
+    }
+
+    public ChestMenu setSize(int size) {
+        if (size % 9 == 0 && size >= 0 && size < 55) {
+            // Resize items list to match actual inventory size in order to reset inventory.
+            // I'm sure that use size of items as inventory size is somehow strange.
+            if (size > items.size()) {
+                while (items.size() < size) {
+                    this.items.add(null);
+                }
+            } else if (size < items.size()) {
+                while (items.size() > size) {
+                    this.items.remove(items.size() - 1);
+                }
+            } else {
+                return this;
+            }
+
+            this.size = size;
+
+            reset(false);
+
+            return this;
+        } else {
+            throw new IllegalArgumentException(
+                    "The size of a ChestMenu must be a multiple of 9 and within the bounds 0-54,"
+                            + " received: "
+                            + size);
+        }
+    }
+
+    public boolean isSizeAutomaticallyInferred() {
+        return size == -1;
     }
 
     @FunctionalInterface
